refactor: reduce long files and functions
Split long functions into smaller helpers and extract modules from oversized files:

- tab-lifecycle.js: split switchTo() into _cleanupPreviousTab() and _activateTab()
- flow-view.js: split _renderList() into _buildGroupParams(), _renderCategorizedGroups(), _renderUncategorizedSection()
- tab-manager.js: extract _setupBusListeners() from init()
- workspace-layout.js (378->147 lines): extract resize/panel logic into workspace-resize.js and serialization into workspace-serializer.js
